pyproject.toml
[project]
name = "llm-wiki-builder"
version = "0.1.0"
description = "Universal LLM-powered wiki builder for any folder — Obsidian-compatible output"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "openai>=1.30.0",
    "click>=8.1",
    "rich>=13.0",
    "pyyaml>=6.0",
    "python-docx>=1.1",
    "openpyxl>=3.1",
    "pdfplumber>=0.10",
    "python-pptx>=0.6",
    "jinja2>=3.1",
]

[project.scripts]
wiki-builder = "wiki_builder.cli:cli"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[tool.setuptools.packages.find]
where = ["."]
include = ["wiki_builder*"]
